Separated the node class into its own file
