+++ src/cds/stack/lockfree_stack.cpp
@@ -1,4 +1,5 @@
 #include "../stack/lockfree_stack.h"
+#include "../stack/node.h"
 #include "../../utility/memory.h"
 
 #include <atomic>
